+++ vercel.json
@@ -1,0 +1,8 @@
+{
+  "version": 2,
+  "builds": [{ "src": "api/index.js", "use": "@vercel/node" }],
+  "routes": [
+    { "src": "/modeltfjs/(.*)", "dest": "api/index.js" },
+    { "src": "/api/(.*)", "dest": "api/index.js" }
+  ]
+}
